requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
langchain==0.1.9
langchain-together==0.0.3.post1
slowapi==0.1.9
jinja2==3.1.3
python-multipart==0.0.9
sse-starlette==1.8.2 